Redefine Document and Node traits to suit arena allocator. Separate Value into separate module.
